refactor tool specs
