+++ llama-index-integrations/tools/llama-index-tools-agentql/llama_index/tools/agentql/__init__.py
@@ -1,4 +1,15 @@
-from llama_index.tools.agentql.base import AgentQLToolSpec
+from llama_index.tools.agentql.extract_web_data_browser.base import (
+    ExtractWebDataBrowserSpec,
+)
+from llama_index.tools.agentql.extract_web_data_rest_api.base import (
+    ExtractWebDataRestAPISpec,
+)
+from llama_index.tools.agentql.get_web_element_browser.base import (
+    GetWebElementBrowserSpec,
+)
 
-
-__all__ = ["AgentQLToolSpec"]
+__all__ = [
+    "ExtractWebDataBrowserSpec",
+    "ExtractWebDataRestAPISpec",
+    "GetWebElementBrowserSpec",
+]
