+++ llama-index-integrations/tools/llama-index-tools-agentql/llama_index/tools/agentql/extract_web_data_rest_api/base.py
@@ -1,0 +1,100 @@
+from typing import Optional
+import os
+
+from llama_index.core.tools.tool_spec.base import BaseToolSpec
+
+from llama_index.tools.agentql.const import (
+    DEFAULT_API_TIMEOUT_SECONDS,
+    DEFAULT_IS_STEALTH_MODE_ENABLED,
+    DEFAULT_WAIT_FOR_PAGE_LOAD_SECONDS,
+    DEFAULT_IS_SCROLL_TO_BOTTOM_ENABLED,
+    DEFAULT_RESPONSE_MODE,
+    DEFAULT_IS_SCREENSHOT_ENABLED,
+)
+from llama_index.tools.agentql.messages import UNSET_API_KEY_ERROR_MESSAGE
+from llama_index.tools.agentql.utils import _aload_data
+
+
+class ExtractWebDataRestAPISpec(BaseToolSpec):
+    """
+    Extract web data from a REST API.
+    """
+
+    spec_functions = [
+        "extract_web_data_with_rest_api",
+    ]
+
+    def __init__(
+        self,
+        api_key: Optional[str] = None,
+        api_timeout: Optional[int] = DEFAULT_API_TIMEOUT_SECONDS,
+        is_stealth_mode_enabled: Optional[bool] = DEFAULT_IS_STEALTH_MODE_ENABLED,
+        wait_for_page_load: Optional[int] = DEFAULT_WAIT_FOR_PAGE_LOAD_SECONDS,
+        is_scroll_to_bottom_enabled: Optional[
+            bool
+        ] = DEFAULT_IS_SCROLL_TO_BOTTOM_ENABLED,
+        mode: Optional[str] = DEFAULT_RESPONSE_MODE,
+        is_screenshot_enabled: Optional[bool] = DEFAULT_IS_SCREENSHOT_ENABLED,
+    ):
+        """
+        Initialize ExtractWebDataRestAPI.
+
+        Args:
+            api_key: AgentQL API key. You can create one at https://dev.agentql.com.
+            api_timeout: The number of seconds to wait for a request before timing out. Defaults to 900.
+
+            is_stealth_mode_enabled: Whether to enable experimental anti-bot evasion strategies. This feature may not work for all websites at all times.
+            Data extraction may take longer to complete with this mode enabled. Defaults to `False`.
+
+            wait_for_page_load: The number of seconds to wait for the page to load before extracting data. Defaults to 0.
+            is_scroll_to_bottom_enabled: Whether to scroll to bottom of the page before extracting data. Defaults to `False`
+
+            mode: 'standard' uses deep data analysis, while 'fast' trades some depth of analysis for speed and is adequate for most usecases.
+            Learn more about the modes in this guide: https://docs.agentql.com/accuracy/standard-mode. Defaults to 'fast'.
+
+            is_screenshot_enabled: Whether to take a screenshot before extracting data. Returned in 'metadata' as a Base64 string. Defaults to `False`
+        """
+        self._api_key = api_key or os.getenv("AGENTQL_API_KEY")
+        if not self._api_key:
+            raise ValueError(UNSET_API_KEY_ERROR_MESSAGE)
+
+        self.api_timeout = api_timeout
+        self.is_stealth_mode_enabled = is_stealth_mode_enabled
+        self.wait_for_page_load = wait_for_page_load
+        self.is_scroll_to_bottom_enabled = is_scroll_to_bottom_enabled
+        self.mode = mode
+        self.is_screenshot_enabled = is_screenshot_enabled
+
+    async def extract_web_data_with_rest_api(
+        self, url: str, query: Optional[str] = None, prompt: Optional[str] = None
+    ) -> dict:
+        """
+        Extracts structured data as JSON from a web page given a URL using either an AgentQL query or a Natural Language description of the data.
+
+        Args:
+            url: Accepts the URL of the public webpage to extract data from.
+            query: Accepts AgentQL query used to extract the data. The query must be enclosed with curly braces `{}`. Either this field or `prompt` field must be provided.
+            prompt: Accepts Natural Language description of the data to extract from the page. If AgentQL query is not specified, always use the `prompt` field. Either this field or `query` field must be provided.
+
+        Returns:
+            dict: The extracted data.
+        """
+        _params = {
+            "wait_for": self.wait_for_page_load,
+            "is_scroll_to_bottom_enabled": self.is_scroll_to_bottom_enabled,
+            "mode": self.mode,
+            "is_screenshot_enabled": self.is_screenshot_enabled,
+        }
+        _metadata = {
+            "experimental_stealth_mode_enabled": self.is_stealth_mode_enabled,
+        }
+
+        return await _aload_data(
+            url=url,
+            query=query,
+            prompt=prompt,
+            params=_params,
+            metadata=_metadata,
+            api_key=self._api_key,
+            timeout=self.api_timeout,
+        )
