+++ llama-index-integrations/tools/llama-index-tools-agentql/llama_index/tools/agentql/messages.py
@@ -6,6 +6,3 @@ QUERY_PROMPT_EXCLUSIVE_ERROR_MESSAGE = (
 )
 UNSET_API_KEY_ERROR_MESSAGE = "No AgentQL API key provided. You can set your API key in code by specifying the `api_key` argument or by setting the `AGENTQL_API_KEY` environment variable. You can create an API key at https://dev.agentql.com."
 UNAUTHORIZED_ERROR_MESSAGE = "Invalid AgentQL API key provided. Please provide a valid API Key. You can create one at https://dev.agentql.com."
-MISSING_BROWSER_ERROR_MESSAGE = (
-    "Browser Instance not found. A browser instance is required to use this tool."
-)
